refactor(mjh/applications): split AddApplicationDialog.tsx (1,070 LOC) into per-step components
Extracts the monolithic dialog into an orchestrator + sub-components:
- add-application-dialog/useAddApplicationFlow.ts  — state-machine hook
- add-application-dialog/dialogHelpers.ts          — pure utility fns
- add-application-dialog/PasteLinkStep.tsx         — URL paste panel
- add-application-dialog/PasteTextStep.tsx         — JD text paste panel
- add-application-dialog/ManualEntryStep.tsx       — company-name combobox
- add-application-dialog/ProcessingStep.tsx        — spinner step
- add-application-dialog/CompanyConfirmStep.tsx    — review + form step

AddApplicationDialog.tsx reduced from 1,070 → 153 LOC (thin orchestrator).
All 9 AddApplicationDialog tests pass unchanged. No state shape changes.
